+++ include/klartraum/onnx/onnx_network.hpp
